+++ .env.example.yaml
@@ -1,7 +1,6 @@
 # firebase
 FIREBASE_PROJECT_ID: ""
-FIREBASE_CALLBACK_URL: ""
-FIREBASE_CALLBACK_SIGNIN_URL: ""
+FIREBASE_CALLBACK_URL: "http://localhost:3000/authentication"
 
 SENDGRID_API_KEY: ""
 SENDGRID_FROM_EMAIL: ""
