+++ email_template.html
@@ -5,7 +5,13 @@
   <head>
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-    <title>Welcome To Clubs</title>
+    <title>
+      {{if .IsSignIn}}
+        Sign in to Clubs
+      {{else}}
+        Welcome to Clubs
+      {{end}}
+    </title>
     <style>
       /* -------------------------------------
           GLOBAL RESETS
