+++ fix_dir.go
@@ -1,0 +1,14 @@
+package emailer
+
+import "os"
+
+// this is so we can use email_template
+// inside of cloud functions
+const gcloudFuncSourceDir = "serverless_function_source_code"
+
+func fixDir() {
+	fileInfo, err := os.Stat(gcloudFuncSourceDir)
+	if err == nil && fileInfo.IsDir() {
+		_ = os.Chdir(gcloudFuncSourceDir)
+	}
+}
